Create step 1
The necessary libraries (tidyverse and ggplot2) were loaded and the dataset student_depression_dataset.csv was imported. After successfully installing tidyverse, I verified that the file was in the working directory using list.files() and then loaded it using read_csv(). Finally, I used glimpse() to examine the structure of the dataset, confirming that it contains 27,901 rows and 18 variables, suitable for the analysis required in the project.